Changed default theme to light, changed the navbar, and refactored the routes on App.jsx to function properly
