+++ frontend/src/App.jsx
@@ -1,14 +1,15 @@
 import { Routes, Route } from "react-router-dom";
 import "./App.css";
 import Header from "./components/Header";
-import NavBar from "./NavBar";
+import NavBar from "./components/NavBar";
+import SignUpLogin from "./pages/SignupLogin";
 
 function App() {
   return (
     <>
       <Header />
       <Routes>
-        <Route></Route>
+        <Route path="/" element={<SignUpLogin />} />
       </Routes>
       <NavBar />
     </>
